Added more format line specifiersi checking. And Fixed the found errors.
git-svn-id: svn://svn.code.sf.net/p/eapidk/code/trunk@65 3723b92f-f4ce-431d-9be9-3c706c148987

+++ apps/EeePProg/Str2Value.c
@@ -631,7 +631,7 @@ ParseAsciiEqu_1(
       *szEndBracket='\0';
       DO(ParseAsciiEqu_cmn(cAS, szStartBracket+1, &CurValue));
       stSubLen=szEndBracket - szStartBracket;
-      EApiSprintfA( szStartBracket, stSubLen+1, "%*lld", stSubLen, CurValue);
+      EApiSprintfA( szStartBracket, stSubLen+1, "%*lld", ((signed int)stSubLen), CurValue);
       *szEndBracket=' ';
     }
     else
